requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
requests
beautifulsoup4
schedule
lxml